moving quitgame to utilties, simplify logic in prepare method of AnswerViewController
