+++ ios-trivia-game/ViewControllers/ResultsViewController.swift
@@ -35,13 +35,6 @@ class ResultsViewController: UIViewController {
 
     @IBAction func onQuitFromResults(_ sender: UIBarButtonItem) {
         // update user_in_game to remove player from user_in_game
-        
-        FirebaseClient.instance.quitGame(complete: {
-            let storyboard = UIStoryboard(name: "Main", bundle: nil)
-            let destination = storyboard.instantiateViewController(withIdentifier: "com.iostriviagame.maintabviewcontroller") as! MainTabViewController
-            destination.selectedIndex = 0
-            destination.navigationController?.isNavigationBarHidden = true
-            self.present(destination, animated: true, completion: nil)
-        }, onError: { (error) in })
+        Utilities.quitGame(controller: self)
     }
 }
